# json: implement Format on jsonEncoded, fused with decoding