Fixes #119542.

This commit implement `Format` directly on the `jsonEncoded` type, writing into the `*bytes.Buffer` at the same time as decoding. It replaces an implementation which decoded the full JSON object first into an intermediate JSON tree and then formatted that JSON tree into a `*bytes.Buffer` before discarding it.

As a result, this avoids the cost of constructing the intermediate JSON objects. It also avoids needing to pass through `jsonString` objects, which previously caused `[]byte->string` re-allocations and memcpys.

The improvement on the new `FormatJSON` benchmark is as follows:
```
name                   old time/op    new time/op    delta
FormatJSON/encoded-10    2.29µs ± 1%    0.95µs ± 1%  -58.55%  (p=0.000 n=10+9)

name                   old alloc/op   new alloc/op   delta
FormatJSON/encoded-10    2.98kB ± 0%    0.85kB ± 0%  -71.51%  (p=0.000 n=10+10)

name                   old allocs/op  new allocs/op  delta
FormatJSON/encoded-10      68.0 ± 0%       3.0 ± 0%  -95.59%  (p=0.000 n=10+10)
```

The test coverage here is pretty good, as indicated by code coverage and by the failures I encountered in `TestJSONRoundTrip` while making this change.

Release note: None