@nvanbenschoten 2024-02-23
I ran a quick test with a [modified version of kv99](https://github.com/nvanbenschoten/cockroach/commit/89e428b8dce2b2d64bc7212ec10a0db9bfe5662d) which uses an 8KB blobs of JSONB data as each row's value.

I compared before yesterday's initial JSON PR (#119539) with after this one. Note that the before binary is running master with Go 1.21, which we know has better GC behavior than earlier versions of Go, so this may be even more impactful if backported to early versions of crdb that use previous versions of Go.

### Top-level results:
- 26% increase in throughput
- 10% reduction in p50 latency
- 24% reduction in p99 latency
- 26% reduction in GC runs

#### Script:
```bash
roachprod create nathan-json --nodes=2 --gce-machine-type='n2-standard-4'
roachprod stage nathan-json:1 cockroach
roachprod put nathan-json:2 ./artifacts/cockroach

roachprod start nathan-json:1
roachprod run nathan-json:2 -- "./cockroach workload init kv --splits=10 {pgurl:1}"
roachprod run nathan-json:2 -- "./cockroach workload run kv --read-percent=99 --ramp=1m --duration=4m {pgurl:1}"

roachprod stop nathan-json:1
roachprod put nathan-json:1 ./artifacts/cockroach
roachprod start nathan-json:1
roachprod run nathan-json:2 -- "./cockroach workload run kv --read-percent=99 --ramp=1m --duration=4m {pgurl:1}"
```

#### Raw results:
```
# Before:
_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__total
  240.0s        0        1588246         6617.7      1.2      1.0      2.2      3.4     27.3  read

_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__total
  240.0s        0          16017           66.7      3.2      2.9      5.0     10.0     50.3  write

_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__result
  240.0s        0        1604263         6684.4      1.2      1.0      2.4      3.7     50.3

# After:
_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__total
  240.0s        0        1994773         8311.6      0.9      0.9      1.6      2.6     37.7  read

_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__total
  240.0s        0          20358           84.8      2.7      2.5      3.9      5.8     23.1  write

_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__result
  240.0s        0        2015131         8396.4      1.0      0.9      1.8      2.8     37.7
```

<img width="987" alt="Screenshot 2024-02-23 at 11 09 12 AM" src="https://github.com/cockroachdb/cockroach/assets/5438456/a1a2a145-6ed5-43a6-b123-bea1311bfbd7">
<img width="987" alt="Screenshot 2024-02-23 at 11 09 16 AM" src="https://github.com/cockroachdb/cockroach/assets/5438456/413687ab-39bc-484d-98b4-7d567b6a237f">
<img width="984" alt="Screenshot 2024-02-23 at 11 09 34 AM" src="https://github.com/cockroachdb/cockroach/assets/5438456/221005ca-bfcb-4ad0-9fa1-6f07859d8a02">

#### CPU profiles:

Before:
<img width="1587" alt="Screenshot 2024-02-23 at 10 59 54 AM" src="https://github.com/cockroachdb/cockroach/assets/5438456/8753dfbb-c641-46bc-8f31-4d4403a03a34">

After:
<img width="1597" alt="Screenshot 2024-02-23 at 11 06 36 AM" src="https://github.com/cockroachdb/cockroach/assets/5438456/dcf8d03a-83a5-4585-b7d4-4d25a498b79b">