@mgartner 2024-02-23
I guess we can't clear the whole buffer, but we can remember the previous size and `buf.Truncate`, right?





<!-- Sent from Reviewable.io -->